requirements.txt
python-docx==0.8.11
flask==3.0.2
tkinter 